pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "bna-market"
version = "2.0.0"
description = "Nashville real estate market analytics application"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "requests==2.31.0",
    "pandas==2.1.4",
    "numpy==1.26.2",
    "plotly==5.18.0",
    "flask==3.0.0",
    "fredapi==0.5.1",
    "python-dotenv==1.0.0",
]

[project.optional-dependencies]
production = [
    "gunicorn>=21.2.0",
]
dev = [
    "pytest>=7.4.3",
    "black>=23.12.1",
    "flake8>=6.1.0",
    "mypy>=1.7.1",
]

[project.scripts]
bna-market = "bna_market.__main__:main"

[tool.setuptools.packages.find]
where = ["."]

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
    \.git
  | \.venv
  | venv
  | logs
  | __pycache__
)/
'''

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--verbose",
    "--cov=.",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-fail-under=80",
]
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "fredapi.*",
    "plotly.*",
]
ignore_missing_imports = true

[tool.flake8]
max-line-length = 100
extend-ignore = ["E203", "W503"]
exclude = [
    ".git",
    "__pycache__",
    ".venv",
    "venv",
    "logs",
    "*.egg-info",
]
per-file-ignores = [
    "__init__.py:F401",
]

[tool.coverage.run]
source = ["."]
omit = [
    "*/tests/*",
    "*/venv/*",
    "*/.venv/*",
    "*/test_*.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
